pyproject.toml
[project]
name = "max-bot-template"
version = "0.1.0"
description = "Max bot template on python"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "maxapi",
    "alembic>=1.13",
    "python-dotenv",
    "sqlalchemy>=2.0",
    "psycopg[binary]>=3.1",
]

[tool.uv.sources]
maxapi = { git = "https://github.com/max-messenger/max-botapi-python.git" }
